file delete added

+++ client/views.py
@@ -174,8 +174,14 @@ def delete_slot(request, slot):
     return HttpResponseRedirect(reverse('client:client_home'))
 
 def delete_file(request, upload):
+    BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
     upload = Uploads.objects.get(id=upload)
-    upload.delete()
+    filepath = BASE_DIR + "".join([c if c != '/' else '\\' for c in upload.link.url[6:]])
+    print(filepath)
+    if os.path.exists(filepath):
+        os.remove(filepath)
+        upload.delete()
+    
     return HttpResponseRedirect(reverse('client:client_home'))
